+++ structure/Common.java
@@ -4,10 +4,10 @@ public class Common {
 
     // 坐标
     public static class Point {
-        public int x;
-        public int y;
+        public float x;
+        public float y;
         public Point() {}
-        public Point(int x, int y) {
+        public Point(float x, float y) {
             this.x = x;
             this.y = y;
         }
@@ -16,10 +16,10 @@ public class Common {
     // 检测框
     public static class Box {
         public Point point; // 左上角原点
-        public int width;
-        public int height;
+        public float width;
+        public float height;
         public Box() {}
-        public Box(Point point, int width, int height) {
+        public Box(Point point, float width, float height) {
             this.point = point;
             this.width = width;
             this.height = height;
